fix(backend): data quality, Sankalpa correctness, and service hardening (B-H2, B-H3, B-M2, B-M4, B-M5, B-M10, B-M11, B-M14, B-M15, B-L6)
- B-H2/B-H3: add rationale comments to V14/V16 migrations and snapshot comment in SessionLogService
- B-M2: cap PlaylistService.listPlaylists at 200 entries via Pageable
- B-M4: replace double ratio with BigDecimal in SankalpaService.toProgressResponse
- B-M5: replace stripTrailingZeros() with setScale(2, HALF_UP) before entity persist
- B-M10: add @Transactional to replaceObservanceEntries as secondary guard
- B-M11: add updatable=false to PlaylistItemEntity.externalId
- B-M14: clamp client-supplied createdAt to server time if skew exceeds tolerance
- B-M15: add ReferenceSeedGuard startup check for reference data seeding
- B-L6: centralize timestamp parsing in SyncRequestSupport; remove duplicated private methods

+++ backend/src/main/java/com/meditation/backend/sankalpa/SankalpaProgressResponse.java
@@ -1,5 +1,6 @@
 package com.meditation.backend.sankalpa;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record SankalpaProgressResponse(
@@ -18,6 +19,6 @@ public record SankalpaProgressResponse(
     int pendingObservanceCount,
     int targetObservanceCount,
     List<SankalpaObservanceDayResponse> observanceDays,
-    double progressRatio
+    BigDecimal progressRatio
 ) {
 }
